styles of the component partners were changed
