+++ src/components/index.tsx
@@ -3,4 +3,4 @@ export { default as Counter } from './counter';
 export { default as Header } from './header';
 export { default as Banner } from './banner';
 export { default as BannerLower } from './bannerLower';
-export { default as Slider } from './slider';
+export { default as Partners } from './partners';
