+++ src/components/partners/partners.test.js
@@ -1,11 +1,10 @@
 import React from 'react';
-import {Slider} from "../index";
-import ReactDom from 'react-dom';
+import {Partners} from "../index";
 import { render, fireEvent, screen} from "@testing-library/react";
 
 
 test('render a title', () => {
-  const {getByText} = render(<Slider />);
+  const {getByText} = render(<Partners />);
   expect(getByText(/Our partners/i)).toBeInTheDocument();
   const buttonRight = screen.getByTestId('buttonRight');
   const buttonLeft = screen.getByTestId('buttonLeft');
@@ -14,11 +13,11 @@ test('render a title', () => {
 });
 
 test('check if user can scroll the slide', () => {
-  const {getByText} = render(<Slider />);
+  const {getByText} = render(<Partners />);
   const onClick = jest.fn();
   const buttonRight = screen.getByTestId('buttonRight');
   const buttonLeft = screen.getByTestId('buttonLeft');
-  render(<Slider onClick={onClick()} />);
+  render(<Partners onClick={onClick()} />);
   fireEvent.click(buttonRight);
   fireEvent.click(buttonLeft);
   expect(onClick).toHaveBeenCalledTimes(1);
