+++ src/components/partners/partnersComponent/partnerComponent.tsx
@@ -1,17 +1,18 @@
 import React from "react";
 
 interface Props {
-  src: any
+  src: string;
 }
 
-const Slide: React.FunctionComponent<Props> = (props)  =>  {
+const PartnerComponent: React.FunctionComponent<Props> = (props)  =>  {
 
   let imgStyles = {
-    width: 300 + "px",
+    maxWidth: 300 + 'px',
     height: "auto",
+    padding: 10 +'px',
   }
     return ( <img src={props.src} alt='imgSlide' style={imgStyles}/>
   );
 }
 
-export default Slide;
+export default PartnerComponent;
