+++ src/pages/home/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styles from './home.module.scss';
-import { Header, Banner, BannerLower, Slider } from '../../components';
+import { Header, Banner, BannerLower, Partners } from '../../components';
 
 
 const Home = () => {
@@ -10,9 +10,8 @@ const Home = () => {
       <Banner />
       <div className={styles.filter}>Filter</div>
       <div className={styles.about}>About</div>
-      <Slider />
       <BannerLower />
-      <div className={styles.partners}>Partners</div>
+      <Partners />
       <div className={styles.footer}>Footer</div>
     </>
   );
